Improve performance by switching to parial search
This updates the chef node searches to be partial searches which should
help serialization and memory usage locally.
